background: Allow configuring the ubuntu accent color
Move the ubuntu colors configuration to a single widget
(CcUbuntuColorsRow) so that it can be easily included in other widgets
with minimal changes.
We keep resources and the widget itself compiled together with the ubuntu
panel, so that there's no need to change multiple panels build scripts,
given that these are per-process anyways.

Helps with LP: #1968213